requirements.txt
streamlit==1.27.0
google-generativeai==0.3.1
python-dotenv==1.0.0